backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
supabase==1.0.3
httpx==0.24.0
pydantic==1.10.7
python-dotenv==1.0.0
openai==0.27.8
bcrypt==4.0.1
email-validator==2.0.0